ccfri 4764 - Moved all closure components into src/components/closure
