+++ frontend/src/components/closure/ClosureConfirmationDialog.vue
